console: Refactor gateway collaborators edit
Move from class components to functional components
